- Added methods to get/set framebuffer base and size - Extended framebuffer to set size when setting adress - Replaced some constants by defines - Removed defined check for ELF32 within arch/mm/virt.h - Added wip logic for virt_map_address - Added code for virt_init

+++ src/kernel/arch/arm/mmio.c
@@ -29,7 +29,7 @@
  * @todo check and revise
  */
 void __attribute__(( optimize( "O0" ) )) mmio_write( uint32_t reg, uint32_t data ) {
-  *( volatile uint32_t* )( (uintptr_t)reg ) = data;
+  *( volatile uint32_t* )( ( uint32_t )reg ) = data;
 }
 
 /**
@@ -41,5 +41,5 @@ void __attribute__(( optimize( "O0" ) )) mmio_write( uint32_t reg, uint32_t data
  * @todo check and revise
  */
 uint32_t __attribute__(( optimize( "O0" ) )) mmio_read( uint32_t reg ) {
-  return *( volatile uint32_t* )( ( uintptr_t )reg );
+  return *( volatile uint32_t* )( ( uint32_t )reg );
 }
